getting ready for pyinstaller

+++ all.py
@@ -1,17 +1,20 @@
-import tkinter
-from tkinter import *
-from tkinter import simpledialog, ttk
-import cv2
-import numpy as np
-from requests_toolbelt.multipart.encoder import MultipartEncoder
 import io
-from PIL import Image, ImageDraw, ImageFont
+import tkinter
 from datetime import datetime, timezone
-from end import *
+from tkinter import *
+from tkinter import simpledialog
+import numpy as np
 import psycopg2.extras
-from dotenv import load_dotenv
+from PIL import Image, ImageDraw, ImageFont
+from requests_toolbelt.multipart.encoder import MultipartEncoder
+from end import *
 
 load_dotenv()
+
+def resource_path(relative_path):
+    if hasattr(sys, '_MEIPASS'):
+        return os.path.join(sys._MEIPASS, relative_path)
+    return os.path.join(os.path.abspath("."), relative_path)
 
 
 def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard_line, offense, defense, league, year, week, regular_post, play_text):
@@ -94,8 +97,8 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                 source_points = [(160, 800), (158, 400), (540, 400), (540, 600)]
             elif "inside":
                 source_points = [(276, 700), (276, 680), (447, 680), (447, 700)]
-            points_image = cv2.imread('images/nfl_20_30.png')
-            field = cv2.imread('images/nfl_20_30.png')
+            points_image = cv2.imread(resource_path('images/nfl_20_30.png'))
+            field = cv2.imread(resource_path('images/nfl_20_30.png'))
         elif 50 > yard_line > 30:
             if hash_or_num == "Hashmark" and offense_l_or_r == "Offense Left":
                 source_points = [(310, 800), (310, 400), (390, 400), (390, 800)]
@@ -116,8 +119,8 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                 source_points = [(160, 800), (158, 400), (540, 400), (540, 600)]
             elif "inside":
                 source_points = [(276, 700), (276, 680), (447, 680), (447, 700)]
-            points_image = cv2.imread('images/nfl_40_g.png')
-            field = cv2.imread('images/nfl_40_g.png')
+            points_image = cv2.imread(resource_path('images/nfl_40_g.png'))
+            field = cv2.imread(resource_path('images/nfl_40_g.png'))
         cur.execute("select rgb1 from nfl_team_colors where team_name = '" + offense + "'")
         home = cur.fetchone()
         home = eval(home[0])[::-1]
@@ -133,8 +136,8 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
 
             source_points = [(160,835),(160,500),(540,500),(540,670)]
 
-        points_image = cv2.imread('images/Figure_1.png')
-        field = cv2.imread('images/Figure_1.png')
+        points_image = cv2.imread(resource_path('images/Figure_1.png'))
+        field = cv2.imread(resource_path('images/Figure_1.png'))
 
         cur.execute("select rgb1 from college_team_colors where team_name = '" + offense + "'")
         home = cur.fetchone()
@@ -169,7 +172,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
 
 
     while cap.isOpened():
-        key = cv2.waitKey(50) & 0xFF
+        key = cv2.waitKey(25) & 0xFF
 
         ret, frame_cap = cap.read()
         if frame_cap is None:
@@ -184,7 +187,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
         (srcwins, src_pts) = srcpointTracker.update(points_image)
         (dstwins, dst_pts) = field_point_tracker.update(frame_cap)
 
-        if len(boxes) == 0:
+        if len(boxes) == 0 and exists == False:
             cv2.imwrite("file.jpg", frame_cap)
             # select the bounding box of the object we want to track (make
             # sure you press ENTER or SPACE after selecting the ROI)
@@ -214,7 +217,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
             buffered = io.BytesIO()
             image.save(buffered, quality=90, format="JPEG")
             # Construct the URL
-            m = MultipartEncoder(fields={'file': (f, buffered.getvalue(), "image/jpeg")})
+            m = MultipartEncoder(fields={'file': (f, buffered.getvalue(), resource_path("image/jpeg"))})
             r = requests.post(url, data=m, headers={'Content-Type': m.content_type})
             # print('post took ' + str(time.time() - start))
             # POST to the API
@@ -231,7 +234,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
 
             for xydata in detect:
                 for second in detections:
-                    if xydata['class'] != second['class']:
+                    if xydata['class'] != second['class'] or xydata['class'] == second['class']:
                         if xydata['x'] - 5 <= second['x'] <= xydata['x'] + 5 or xydata['x'] == second['x']:
                             if xydata['y'] - 5 <= second['y'] <= xydata['y'] + 5 or xydata['y'] == second['y']:
                                 second['remove'] = 'yes'
@@ -300,12 +303,13 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                             np.split(points_out, 1)
                             for x in points_out:
                                 x1, y1 = x[0]
-                                cv2.circle(field, (int(abs(x1) + 50), int(abs(y1))), 5, (0, 0, 0), 2)
-                                cur.execute(
-                                    sql.SQL("INSERT INTO main_table VALUES (%s, %s, %s, %s, %s, %s, %s, %s)"),
-                                    (gameid_number, str(playid_number), int(box_id), (int(abs(x1))), (int(abs(y1))),
-                                         cap.get(cv2.CAP_PROP_POS_FRAMES), null_variable,
-                                         str(face_no)))
+                                cv2.rectangle(field, ((int(abs(x1))), int(abs(y1))),
+                                              (int(abs(x1)) + 1, int(abs(y1)) + 1), (0, 0, 0), 2)
+                                tup = (gameid_number, str(playid_number), int(box_id), (int(abs(x1))), (int(abs(y1))),
+                                       cap.get(cv2.CAP_PROP_POS_FRAMES), null_variable,
+                                       null_variable, gameid_number)
+
+                                biglist.append(tup)
                                 screenshot += 1
                                 break
                             speed_pts2.clear()
@@ -317,8 +321,12 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                 (x, y, w, h) = [int(v) for v in box]
                 pt3.append(((x + int(w / 2), y + h), face_no))
                 cv2.rectangle(frame_cap, (x, y), (x + w, y + h), (0, 255, 0), 3)
-                cv2.putText(frame_cap, str((json_prediction[face_no]) + " " + str(face_no)), (x, y - 30), cv2.FONT_HERSHEY_TRIPLEX,
+                try:
+                    cv2.putText(frame_cap, str((json_prediction[face_no]) + " " + str(face_no)), (x, y - 30), cv2.FONT_HERSHEY_TRIPLEX,
                             .7, (0, 255, 0), 1, cv2.LINE_AA)
+                except (IndexError):
+                    cv2.putText(frame_cap, str(face_no + total_players), (x, y - 30), cv2.FONT_HERSHEY_TRIPLEX,
+                                .7, (0, 0, 0), 1, cv2.LINE_AA)
 
                 if len(source_points) > 0:
                     if len(field_points) > 0:
@@ -364,7 +372,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
         cv2.imshow("Frame", frame_cap)
         if screenshot > 0 or exists == True:
             cv2.imshow("field", field)
-            cv2.destroyWindow("points_image")
+            cv2.destroyWindow("points")
 
         # if the 'space' key is selected, we are going to "select" a bounding
         # box to track
@@ -469,6 +477,7 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                             (gameid_number, playid_number, offense, defense,
                              play_text, datetime.now(timezone.utc), league, year, regular_post, week))
 
+
             cur.execute(
                 "select frame, position, color, y_pos, playerid from main_table where playid = '" + playid_number + "' order by frame asc")
             off_vs_def = cur.fetchall()
@@ -509,18 +518,39 @@ def opencv(file, hash_or_num, gameid_number, playid_number, offense_l_or_r, yard
                                     "update main_table set color = 'offense' where playid = '" + playid_number + "' and playerid = %s"),
                                             (playerid,))
 
-                cur.execute("select distinct(playerid), position from main_table where playid = '" + playid_number + "';")
-                no_position = cur.fetchall()
-                if len(no_position) > 0:
-                    for p in no_position:
-                        if p[1] == None or p[1] == "":
-                            USER_INP = simpledialog.askstring(title="Player Identification",
-                                                              prompt="What Position is the Player Id?" + str(p[1]))
-                            cur.execute(sql.SQL("update {} set position = %s where playerid = %s and playid = %s;").format(
-                                sql.Identifier('main_table')),
-                                (USER_INP, int(p[0]), str(playid_number)))
-                else:
-                    print("all positions updated")
+            # get players with no position and update them
+            cur.execute("select distinct(playerid), position from main_table where playid = '" + playid_number + "';")
+            no_position = cur.fetchall()
+            if len(no_position) > 0:
+                for p in no_position:
+                    if p[1] == None or p[1] == "":
+                        USER_INP = simpledialog.askstring(title="Player Identification",
+                                                          prompt="What Position is the Player Id?" + str(p[1]))
+                        cur.execute(sql.SQL("update {} set position = %s where playerid = %s and playid = %s;").format(
+                            sql.Identifier('main_table')),
+                            (USER_INP, int(p[0]), str(playid_number)))
+
+                for x in off_vs_def:
+                    if x[4] == 'C':
+                        cur.execute(sql.SQL(
+                            "update main_table set position = 'C' where playid = '" + playid_number + "' and playerid = %s"),
+                                    (int(USER_INP),))
+                        y_pos = x[3]
+                        for y in off_vs_def:
+                            playerid = y[4]
+                            if y[3] < y_pos:
+                                cur.execute(sql.SQL(
+                                    "update main_table set color = 'defense' where playid = '" + playid_number + "' and playerid = %s"),
+                                            (playerid,))
+                            else:
+                                cur.execute(sql.SQL(
+                                    "update main_table set color = 'offense' where playid = '" + playid_number + "' and playerid = %s"),
+                                            (playerid,))
+
+            else:
+                print("all positions updated")
+
+
 
             cur.close()
             conn.close()
